move from core to cli utils
