+++ packages/cli/src/commands/test.ts
@@ -1,25 +1,25 @@
 import { loadProject } from '@/src/project';
 import { AgentServer } from '@/src/server/index';
 import { jsonToCharacter, loadCharacterTryPath } from '@/src/server/loader';
-import { TestRunner, buildProject, promptForEnvVars } from '@/src/utils';
 import {
-  type IAgentRuntime,
-  type ProjectAgent,
-  resolvePgliteDir,
+  TestRunner,
+  buildProject,
+  promptForEnvVars,
   resolveEnvFile,
-} from '@elizaos/core';
+  resolvePgliteDir,
+} from '@/src/utils';
+import { type IAgentRuntime, type ProjectAgent } from '@elizaos/core';
 import { Command, Option } from 'commander';
 import * as dotenv from 'dotenv';
+import { exec } from 'node:child_process';
 import * as fs from 'node:fs';
 import { existsSync } from 'node:fs';
 import * as net from 'node:net';
 import * as os from 'node:os';
 import path from 'node:path';
-import { exec } from 'node:child_process';
 import { promisify } from 'node:util';
-import { startAgent } from './start';
 import { pathToFileURL } from 'url';
-
+import { startAgent } from './start';
 const execAsync = promisify(exec);
 
 // Helper function to check port availability
