+++ packages/cli/src/utils/index.ts
@@ -6,6 +6,7 @@ export * from './config-manager';
 export * from './copy-template';
 export * from './display-banner';
 export * from './env-prompt';
+export * from './get-config';
 export * from './get-package-info';
 export * from './github';
 export * from './handle-error';
@@ -16,8 +17,7 @@ export * from './package-manager';
 export * from './port-handling';
 export * from './publisher';
 export * from './resolve-import';
+export * from './resolve-utils';
 export * from './run-bun';
 export * from './testRunner';
 export * from './user-environment';
-export * from './get-config';
-export * from './env-utils';
